// carros/client/src/app/models/employee.model.ts
export class Employee {
    id: number;
    naem: string;
    birth: Date;
    gender: string;
    cpf: string;
    number: string;
    street: string;
    district: string;
    city: string;
    state: string;
    cep: string;
    salary: string;
    situation: string;
    role: number;
    branch_id: number;
}